requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24
tqdm>=4.0
streamlit>=1.27.0
python-dotenv>=1.0.0
PyPDF2>=3.0
flask==2.3.2
